chore(mc): Refactor TopSites tests to account for separate files (nobug)
