+++ system-addon/content-src/components/TopSites/TopSitesConstants.js
@@ -1,0 +1,5 @@
+module.exports = {
+  TOP_SITES_SOURCE: "TOP_SITES",
+  TOP_SITES_CONTEXT_MENU_OPTIONS: ["CheckPinTopSite", "Separator", "OpenInNewWindow",
+    "OpenInPrivateWindow", "Separator", "BlockUrl", "DeleteUrl"]
+};
